pyproject.toml
[project]
name = "claude-hub"
version = "0.1.0"
description = "Visual dashboard for managing Claude Code harness configuration"
requires-python = ">=3.13"
license = {text = "MIT"}
readme = "README.md"
keywords = ["claude", "claude-code", "dashboard", "configuration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: User Interfaces",
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-frontmatter>=1.1",
    "pydantic>=2.10",
    "httpx>=0.28",
    "aiosqlite>=0.22.1",
    "pywebview>=6.1",
]

[project.urls]
Homepage = "https://github.com/claude-hub/claude-hub"
Repository = "https://github.com/claude-hub/claude-hub"

[project.scripts]
claude-hub = "claude_hub.main:cli"
claude-hub-tracker = "claude_hub.tracker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
artifacts = [
    "src/claude_hub/static/**",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/claude_hub"]


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]
